Rename linking functions and update documentation for #18
