+++ gslab_make/create_inputs.py
@@ -40,7 +40,7 @@ def create_links(paths,
     move_dir = paths['move_dir']
 
     try:              
-        move_list = MoveList(file_list, file_format, move_dir, mapping_dict)
+        move_list = MoveList(file_list, move_dir, mapping_dict)
         if move_list.move_directive_list:
             os.makedirs(move_dir)
             move_map = move_list.create_symlinks()       
@@ -50,12 +50,8 @@ def create_links(paths,
         write_to_makelog(paths, 'Links successfully created!')    
         return(move_map)
     except:
-        error_message = 'An error was encountered with `create_links`. Traceback can be found below.' 
-        error_message = format_error(error_message) + '\n' + traceback.format_exc()
-        write_to_makelog(paths, error_message)
-        
-        raise               
-
+        raise CritError(traceback.format_exc())
+		
 
 def copy_inputs(paths,
                 file_list,
@@ -82,19 +78,20 @@ def copy_inputs(paths,
     move_map : list
         List of (source, destination) for each copy created.
     """
-
-    link_dir = paths['input_dir']
+		
+    move_dir = paths['input_dir']
 
     try:              
-        link_list = LinksList(file_list, "input", link_dir, mapping_dict, True)
-        if link_list.link_directive_list:
-            os.makedirs(link_dir)
-            link_map = link_list.create_symlinks()       
+        move_list = MoveList(file_list, move_dir, mapping_dict)
+        if move_list.move_directive_list:
+            os.makedirs(move_dir)
+            move_map = move_list.create_copies()       
         else:
-            link_map = []
+            move_map = []
 
         write_to_makelog(paths, 'Copies successfully created!')    
-        return(link_map)
+        return(move_map)
+		
     except:
         error_message = 'An error was encountered with `copy_inputs`. Traceback can be found below.' 
         error_message = format_error(error_message) + '\n' + traceback.format_exc()
@@ -129,10 +126,16 @@ def link_inputs(paths,
     move_map : list
         List of (source, destination) for each symlink created.
     """
-    paths['move_dir'] = paths['input_dir']
-    move_map = create_links(paths, file_list, mapping_dict)
-    return(move_map)
-
+    try:
+		paths['move_dir'] = paths['input_dir']
+		move_map = create_links(paths, file_list, mapping_dict)
+		return(move_map)
+	except:
+        error_message = 'An error was encountered with `link_inputs`. Traceback can be found below.' 
+        error_message = format_error(error_message) + '\n' + traceback.format_exc()
+        write_to_makelog(paths, error_message)
+        
+        raise 
 
 def link_externals(paths,
                    file_list,
@@ -160,6 +163,14 @@ def link_externals(paths,
     move_map : list
         List of (source, destination) for each symlink created.
     """
-    paths['move_dir'] = paths['external_dir']
-    move_map = create_links(paths, file_list, mapping_dict)
-    return(move_map)
+    
+	try:
+		paths['move_dir'] = paths['external_dir']
+		move_map = create_links(paths, file_list, mapping_dict)
+		return(move_map)
+	except:
+        error_message = 'An error was encountered with `link_externals`. Traceback can be found below.' 
+        error_message = format_error(error_message) + '\n' + traceback.format_exc()
+        write_to_makelog(paths, error_message)
+        
+        raise
